Add list planets and find by id endpoints

+++ src/main/java/com/grglucastr/universeapi/models/Planet.java
@@ -9,7 +9,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -22,5 +21,5 @@ public class Planet{
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private BigDecimal mass;
+    private Double mass;
 }
